normaliz: build in release mode
This sets NDEBUG, which in turn disables asserts, which in turn avoids
hardcoding the PREFIX path for debug messages. In the future (for
"shipping" versions of the bundle), we may wish to re-enable this again,
as those asserts could be useful. But for now, it helps to have these
gone.